requirements copy.txt
# Automatically generated by https://github.com/damnever/pigar.

bunnet==1.3.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-SocketIO==5.3.6
gunicorn==22.0.0
numpy==1.26.4
pandas==2.2.2
pandas_ta==0.3.14b0
pydantic==2.7.1
PyJWT==2.8.0
pymongo==4.7.2
python-dotenv==1.0.1
python-okx==0.2.9
requests==2.31.0
python-binance==1.0.19
eventlet